Move MockMotorsBus
